task_manager/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.1
sqlparse==0.4.4
pytz==2023.3
